add support for backtester to generate an HTML document with some basic visualizations
