+++ scripts/backtest_plugins.ts
@@ -1,116 +1,13 @@
 import { BigNumber, providers } from 'ethers'
-import { BytesLike, formatEther } from 'ethers/lib/utils'
+import { formatEther } from 'ethers/lib/utils'
 import { ethers } from 'hardhat'
 import { bn, fp } from '#/common/numbers'
 import { networkConfig } from '#/common/configuration'
 import { oracleTimeout } from './deployment/utils'
-import fetch from 'isomorphic-fetch'
 import fs from 'fs'
+import { backTestPlugin } from './backtester/backtester'
 
-const sleep = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms))
-
-export const submitBacktest = async (
-  backtestServiceUrl: string,
-  deploymentTransactionData: BytesLike,
-  start: number,
-  stride: number,
-  numberOfSamples: number
-) => {
-  const resp = await fetch(`${backtestServiceUrl}/api/backtest-plugin`, {
-    method: 'POST',
-    body: JSON.stringify({
-      byteCode: deploymentTransactionData,
-      stride,
-      startBlock: start,
-      samples: numberOfSamples,
-    }),
-    headers: {
-      'Content-Type': 'application/json',
-    },
-  })
-  let result: any = await resp.json()
-
-  return result
-}
-
-export const awaitBacktestJobResult = async (backtestServiceUrl: string, key: string) => {
-  let result: any
-  while (1) {
-    await sleep(2000)
-    result = await (
-      await fetch(`${backtestServiceUrl}/api/backtest-plugin-status/${key}`, {
-        method: 'GET',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-      })
-    ).json()
-
-    if (result.jobStatus !== 'RUNNING') {
-      return await (
-        await fetch(`${backtestServiceUrl}/api/backtest-plugin/${key}`, {
-          method: 'GET',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-        })
-      ).json()
-    }
-  }
-}
-
-const backTestPlugin = async <TT, T>(
-  parametersToTest: { variantName: string; config: TT; additionalArgs: T }[],
-  createDeployTx: (testParams: { config: TT; additionalArgs: T }) => Promise<BytesLike>,
-  opts: {
-    start: number
-    stride: number
-    numberOfSamples: number
-    backtestServiceUrl: string
-  }
-) => {
-  return await Promise.all(
-    parametersToTest.map(async (params) => {
-      try {
-        const deployTx = await createDeployTx(params)
-        const backtestJob = await submitBacktest(
-          opts.backtestServiceUrl,
-          deployTx,
-          opts.start,
-          opts.stride,
-          opts.numberOfSamples
-        )
-
-        const backtestJobResult = await awaitBacktestJobResult(
-          opts.backtestServiceUrl,
-          backtestJob.hash
-        )
-
-        return {
-          status: backtestJobResult.jobStatus,
-          backtestName: params.variantName,
-          constructorArgs: {
-            config: params.config,
-            additionalArgs: params.additionalArgs,
-          },
-          result: backtestJobResult,
-        }
-      } catch (e: any) {
-        console.error(`Failed to ${e} run backtest for ${params.variantName}`)
-        console.log('Skking')
-        return {
-          status: 'FAILED',
-          backtestName: params.variantName,
-          constructorArgs: {
-            config: params.config,
-            additionalArgs: params.additionalArgs,
-          },
-          error: e.toString(),
-        }
-      }
-    })
-  )
-}
+const htmlReportTemplate = fs.readFileSync("./scripts/backtester/report-template.html", "utf8")
 
 export const main = async () => {
   const provider = new providers.JsonRpcProvider(process.env.MAINNET_RPC_URL)
@@ -186,6 +83,11 @@ export const main = async () => {
     fs.writeFileSync(
       `${process.env.BACKTEST_RESULT_DIR}/overview.json`,
       JSON.stringify(overview, null, 2)
+    )
+    const htmlReport = htmlReportTemplate.replace("const data = []", "const data = " + JSON.stringify(backTests, null, 2))
+    fs.writeFileSync(
+      `${process.env.BACKTEST_RESULT_DIR}/report.html`,
+      htmlReport
     )
   } else {
     const result = {
